pyproject.toml
[tool.poetry]
name = "scholar-ai"
version = "0.1.0"
description = "ScholarAI Backend System with enhanced WebSearch Agent and B2 integration"
authors = ["ScholarAI Team"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
pydantic = "^2.11.3"
aio-pika = "^9.4.3"
pika = "^1.3.2"
python-dotenv = "^1.1.0"
# Enhanced WebSearch Agent Dependencies
httpx = "^0.27.0"
aiohttp = "^3.9.1"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
PyPDF2 = "^3.0.1"
pdfplumber = "^0.10.0"
google-generativeai = "^0.8.3"
numpy = "^1.24.3"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
seaborn = "^0.12.2"
plotly = "^5.18.0"
scikit-learn = "^1.3.2"
nltk = "^3.8.1"
spacy = "^3.7.2"
textblob = "^0.17.1"
wordcloud = "^1.9.2"
feedparser = "^6.0.10"
crossref-commons = "^0.0.7"
pytest-asyncio = "^1.0.0"
# Backblaze B2 for PDF storage
b2sdk = "^2.5.0"
# Text extraction dependencies
pytesseract = "^0.3.10"
Pillow = "^10.0.0"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
flake8 = "^7.2.0"
pytest = "^8.3.5"
blis = "^0.7.11"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

